+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/abstracts/PossiblyMediaGroupMessage.kt
@@ -1,8 +1,8 @@
 package dev.inmo.tgbotapi.types.message.abstracts
 
-import dev.inmo.tgbotapi.types.MediaGroupIdentifier
+import dev.inmo.tgbotapi.types.MediaGroupId
 import dev.inmo.tgbotapi.types.message.content.MessageContent
 
 interface PossiblyMediaGroupMessage<out T : MessageContent> : ContentMessage<T> {
-    val mediaGroupId: MediaGroupIdentifier?
+    val mediaGroupId: MediaGroupId?
 }