+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/queries/callback/InlineMessageIdGameShortNameCallbackQuery.kt
@@ -4,7 +4,7 @@ import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.chat.CommonUser
 
 data class InlineMessageIdGameShortNameCallbackQuery(
-    override val id: CallbackQueryIdentifier,
+    override val id: CallbackQueryId,
     override val from: CommonUser,
     override val chatInstance: String,
     override val inlineMessageId: InlineMessageId,